Fix Gcc-11/C++17 warnings

+++ src/meshTools/triSurface/triSurface/interfaces/STL/readSTLASCII.C
@@ -15,7 +15,7 @@
 #define FLEX_BETA
 #endif
 
- 
+
 /* First, we deal with  platform-specific or compiler-specific issues. */
 
 /* begin standard C headers. */
@@ -32,7 +32,7 @@
 #if defined (__STDC_VERSION__) && __STDC_VERSION__ >= 199901L
 
 /* C99 says to define __STDC_LIMIT_MACROS before including stdint.h,
- * if you want the limit (max/min) macros for int types. 
+ * if you want the limit (max/min) macros for int types.
  */
 #ifndef __STDC_LIMIT_MACROS
 #define __STDC_LIMIT_MACROS 1
@@ -49,7 +49,7 @@ typedef uint32_t flex_uint32_t;
 typedef signed char flex_int8_t;
 typedef short int flex_int16_t;
 typedef int flex_int32_t;
-typedef unsigned char flex_uint8_t; 
+typedef unsigned char flex_uint8_t;
 typedef unsigned short int flex_uint16_t;
 typedef unsigned int flex_uint32_t;
 
@@ -87,21 +87,21 @@ typedef unsigned int flex_uint32_t;
 #endif /* ! FLEXINT_H */
 
 /* begin standard C++ headers. */
-#include <iostream> 
+#include <iostream>
 #include <errno.h>
 #include <cstdlib>
 #include <cstdio>
 #include <cstring>
 /* end standard C++ headers. */
 
- 
+
 /* Promotes a possibly negative, possibly signed char to an unsigned
  * integer for use as an array index.  If the signed char is negative,
  * we want to instead treat it as an 8-bit unsigned char, hence the
  * double cast.
  */
 #define YY_SC_TO_UI(c) ((unsigned int) (unsigned char) c)
- 
+
 /* Translate the current start state into a value that can be later handed
  * to BEGIN to return to the state.  The YYSTATE alias is for lex
  * compatibility.
@@ -114,7 +114,7 @@ typedef unsigned int flex_uint32_t;
 /* Special action meaning "start processing a new file". */
 #define YY_NEW_FILE yyrestart( yyin  )
 
- 
+
 /* Size of default input buffer. */
 #ifndef YY_BUF_SIZE
 #ifdef __ia64__
@@ -127,13 +127,13 @@ typedef unsigned int flex_uint32_t;
 #define YY_BUF_SIZE 16384
 #endif /* __ia64__ */
 #endif
- 
- 
+
+
 #ifndef YY_TYPEDEF_YY_BUFFER_STATE
 #define YY_TYPEDEF_YY_BUFFER_STATE
 typedef struct yy_buffer_state *YY_BUFFER_STATE;
 #endif
- 
+
 
 extern size_t yyleng;
 
@@ -148,7 +148,7 @@ extern size_t yyleng;
 struct yy_buffer_state
 	{
 
-	std::streambuf* yy_input_file; 
+	std::streambuf* yy_input_file;
 
 	char *yy_ch_buf;		/* input buffer */
 	char *yy_buf_pos;		/* current position in input buffer */
@@ -184,7 +184,7 @@ struct yy_buffer_state
 
     int yy_bs_lineno; /**< The line count. */
     int yy_bs_column; /**< The column count. */
-    
+
 	/* Whether to try to fill the input buffer when we reach the
 	 * end of it.
 	 */
@@ -226,7 +226,7 @@ struct yy_buffer_state
 
 void *yyalloc (size_t  );
 void *yyrealloc (void *,size_t  );
- 
+
 
 /* Begin user sect3 */
 
@@ -821,7 +821,7 @@ static const flex_int16_t yy_chk[594] =
       277,  277,  277
 
     } ;
- 
+
 #define YY_RESTORE_YY_MORE_OFFSET
 // !Line1 "readSTLASCII-.L"
 /*--------------------------------*- C++ -*----------------------------------*\
@@ -997,8 +997,8 @@ STLLexer::STLLexer(std::istream* is, const label approxNpoints)
  */
 #include <unistd.h>
 #endif
- 
- 
+
+
 
 /* Amount of stuff to slurp up with each read. */
 #ifndef YY_READ_BUF_SIZE
@@ -1056,7 +1056,7 @@ STLLexer::STLLexer(std::istream* is, const label approxNpoints)
 #ifndef YY_USER_ACTION
 #define YY_USER_ACTION
 #endif
- 
+
 
 #define  \
 	YY_USER_ACTION
@@ -1068,7 +1068,7 @@ YY_DECL
 	yy_state_type yy_current_state;
 	char *yy_cp, *yy_bp;
 	int yy_act;
-    
+
 	if ( !(yy_init) )
 		{
 		(yy_init) = 1;
@@ -1633,7 +1633,7 @@ int yyFlexLexer::LexerInput( char* buf, int max_size )
 
 void yyFlexLexer::LexerOutput( const char* buf, int size )
 {
-	
+
 }
 
 /* yy_get_next_buffer - try to read in a new buffer
@@ -1756,7 +1756,7 @@ int yyFlexLexer::yy_get_next_buffer()
 	else
 		ret_val = EOB_ACT_CONTINUE_SCAN;
 
-	if ((int) ((yy_n_chars) + number_to_move) > YY_CURRENT_BUFFER_LVALUE->yy_buf_size) {
+	if ( ((yy_n_chars) + number_to_move) > YY_CURRENT_BUFFER_LVALUE->yy_buf_size) {
 		/* Extend the array by 50%, plus the number we really need. */
 		int new_size = (yy_n_chars) + number_to_move + ((yy_n_chars) >> 1);
 		YY_CURRENT_BUFFER_LVALUE->yy_ch_buf = (char *) yyrealloc((void *) YY_CURRENT_BUFFER_LVALUE->yy_ch_buf,new_size  );
@@ -1779,7 +1779,7 @@ int yyFlexLexer::yy_get_next_buffer()
 {
 	yy_state_type yy_current_state;
 	char *yy_cp;
-    
+
 	yy_current_state = (yy_start);
 
 	for ( yy_cp = (yytext_ptr) + 0; yy_cp < (yy_c_buf_p); ++yy_cp )
@@ -1834,7 +1834,7 @@ int yyFlexLexer::yy_get_next_buffer()
     void yyFlexLexer::yyunput( int c, char* yy_bp)
 {
 	char *yy_cp;
-    
+
     yy_cp = (yy_c_buf_p);
 
 	/* undo effects of setting up yytext */
@@ -1872,7 +1872,7 @@ int yyFlexLexer::yy_get_next_buffer()
     int yyFlexLexer::yyinput()
 {
 	int c;
-    
+
 	*(yy_c_buf_p) = (yy_hold_char);
 
 	if ( *(yy_c_buf_p) == 0 )
@@ -1938,12 +1938,12 @@ int yyFlexLexer::yy_get_next_buffer()
 
 /** Immediately switch to a different input stream.
  * @param input_file A readable stream.
- * 
+ *
  * @note This function does not reset the start condition to @c INITIAL .
  */
     void yyFlexLexer::yyrestart( std::istream& input_file )
 {
-    
+
 	if ( ! YY_CURRENT_BUFFER ){
         yyensure_buffer_stack ();
 		YY_CURRENT_BUFFER_LVALUE =
@@ -1956,7 +1956,7 @@ int yyFlexLexer::yy_get_next_buffer()
 
 /** Delegate to the new version that takes an istream reference.
  * @param input_file A readable stream.
- * 
+ *
  * @note This function does not reset the start condition to @c INITIAL .
  */
 void yyFlexLexer::yyrestart( std::istream* input_file )
@@ -1966,11 +1966,11 @@ void yyFlexLexer::yyrestart( std::istream* input_file )
 
 /** Switch to a different input buffer.
  * @param new_buffer The new input buffer.
- * 
+ *
  */
     void yyFlexLexer::yy_switch_to_buffer( YY_BUFFER_STATE new_buffer )
 {
-    
+
 	/* TODO. We should be able to replace this entire function body
 	 * with
 	 *		yypop_buffer_state();
@@ -2010,13 +2010,13 @@ void yyFlexLexer::yyrestart( std::istream* input_file )
 /** Allocate and initialize an input buffer state.
  * @param file A readable stream.
  * @param size The character buffer size in bytes. When in doubt, use @c YY_BUF_SIZE.
- * 
+ *
  * @return the allocated buffer state.
  */
     YY_BUFFER_STATE yyFlexLexer::yy_create_buffer( std::istream& file, int size )
 {
 	YY_BUFFER_STATE b;
-    
+
 	b = (YY_BUFFER_STATE) yyalloc(sizeof( struct yy_buffer_state )  );
 	if ( ! b )
 		YY_FATAL_ERROR( "out of dynamic memory in yy_create_buffer()" );
@@ -2040,7 +2040,7 @@ void yyFlexLexer::yyrestart( std::istream* input_file )
 /** Delegate creation of buffers to the new version that takes an istream reference.
  * @param file A readable stream.
  * @param size The character buffer size in bytes. When in doubt, use @c YY_BUF_SIZE.
- * 
+ *
  * @return the allocated buffer state.
  */
 	YY_BUFFER_STATE yyFlexLexer::yy_create_buffer( std::istream* file, int size )
@@ -2050,11 +2050,11 @@ void yyFlexLexer::yyrestart( std::istream* input_file )
 
 /** Destroy the buffer.
  * @param b a buffer created with yy_create_buffer()
- * 
+ *
  */
     void yyFlexLexer::yy_delete_buffer( YY_BUFFER_STATE b )
 {
-    
+
 	if ( ! b )
 		return;
 
@@ -2075,7 +2075,7 @@ void yyFlexLexer::yyrestart( std::istream* input_file )
 
 {
 	int oerrno = errno;
-    
+
 	yy_flush_buffer( b );
 
 	b->yy_input_file = file.rdbuf();
@@ -2096,7 +2096,7 @@ void yyFlexLexer::yyrestart( std::istream* input_file )
 
 /** Discard all buffered characters. On the next scan, YY_INPUT will be called.
  * @param b the buffer state to be flushed, usually @c YY_CURRENT_BUFFER.
- * 
+ *
  */
     void yyFlexLexer::yy_flush_buffer( YY_BUFFER_STATE b )
 {
@@ -2125,7 +2125,7 @@ void yyFlexLexer::yyrestart( std::istream* input_file )
  *  the current state. This function will allocate the stack
  *  if necessary.
  *  @param new_buffer The new state.
- *  
+ *
  */
 void yyFlexLexer::yypush_buffer_state (YY_BUFFER_STATE new_buffer)
 {
@@ -2155,7 +2155,7 @@ void yyFlexLexer::yypush_buffer_state (YY_BUFFER_STATE new_buffer)
 
 /** Removes and deletes the top of the stack, if present.
  *  The next element becomes the new top.
- *  
+ *
  */
 void yyFlexLexer::yypop_buffer_state (void)
 {
@@ -2179,7 +2179,7 @@ void yyFlexLexer::yypop_buffer_state (void)
 void yyFlexLexer::yyensure_buffer_stack(void)
 {
 	size_t num_to_alloc;
-    
+
 	if (!(yy_buffer_stack)) {
 
 		/* First allocation is just for 2 elements, since we don't know if this
@@ -2192,9 +2192,9 @@ void yyFlexLexer::yyensure_buffer_stack(void)
 								);
 		if ( ! (yy_buffer_stack) )
 			YY_FATAL_ERROR( "out of dynamic memory in yyensure_buffer_stack()" );
-								  
+
 		memset((yy_buffer_stack), 0, num_to_alloc * sizeof(struct yy_buffer_state*));
-				
+
 		(yy_buffer_stack_max) = num_to_alloc;
 		(yy_buffer_stack_top) = 0;
 		return;
@@ -2265,7 +2265,7 @@ void yyFlexLexer::LexerError( const char msg[] )
 }
 
 
- 
+
 
 void *yyalloc (size_t  size )
 {
@@ -2274,7 +2274,7 @@ void *yyalloc (size_t  size )
 
 void *yyrealloc  (void * ptr, size_t  size )
 {
-		
+
 	/* The cast to (char *) in the following accommodates both
 	 * implementations that use char* generic pointers, and those
 	 * that use void* generic pointers.  It works with the latter
@@ -2378,4 +2378,3 @@ bool triSurface::readSTLASCII(const fileName& STLfileName)
  /* ------------------------------------------------------------------------ *\
     ------ End of STLToFoam.L
  \* ------------------------------------------------------------------------ */
-
